pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ask-my-docs"
version = "1.0.0"
description = "Production RAG application with hybrid retrieval, cross-encoder reranking, and citation enforcement"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Muhammad Farooq", email = "mfarooqshafee333@gmail.com" },
]
keywords = ["rag", "retrieval", "nlp", "llm", "fastapi", "faiss", "bm25"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "rank-bm25>=0.2.2",
    "openai>=1.35.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "pypdf>=4.2.0",
    "beautifulsoup4>=4.12.0",
    "streamlit>=1.36.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/askdocs"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
